add assign/unassign with authorization change display based on logged user

+++ device-management-ui/src/app/devices/device-list.component/device-list.component.ts
@@ -7,6 +7,7 @@ import { Device, DeviceTypeLabel } from '../../core/models/device.model';
 import { Router } from '@angular/router';
 import { ConfirmDialogComponent } from '../../shared/components/confirm-dialog.component/confirm-dialog.component';
 import { MatIconModule } from '@angular/material/icon';
+import { AuthService } from '../../core/services/auth.service';
 
 @Component({
   selector: 'app-device-list.component',
@@ -17,6 +18,7 @@ import { MatIconModule } from '@angular/material/icon';
 export class DeviceListComponent implements OnInit {
   protected readonly deviceService = inject(DeviceService);
   protected readonly router = inject(Router);
+  protected readonly authService = inject(AuthService);
   private readonly toast = inject(ToastService);
 
   protected readonly DeviceTypeLabel = DeviceTypeLabel;
@@ -24,6 +26,10 @@ export class DeviceListComponent implements OnInit {
   protected readonly activeFilter = signal<'all' | 'phone' | 'tablet' | 'unassigned'>('all');
   protected readonly showDeleteDialog = signal(false);
   protected readonly deviceToDelete = signal<Device | null>(null);
+
+  protected readonly message = this.authService.isAdmin()
+    ? 'This device will be permanently removed from the system.'
+    : 'This device will be unassigned from you.';
 
   protected readonly filteredDevices = computed(() => {
     let list = this.deviceService.devices();
@@ -46,8 +52,18 @@ export class DeviceListComponent implements OnInit {
     return list;
   });
 
+  readonly currentUserId = computed(() => this.authService.linkedUserId());
+
   ngOnInit(): void {
     this.deviceService.loadAll().subscribe();
+  }
+
+  isAssignedToMe(assignedUserId?: number | null): boolean {
+    var isAssignedToMe =
+      assignedUserId !== null &&
+      assignedUserId !== undefined &&
+      assignedUserId === this.currentUserId();
+    return isAssignedToMe;
   }
 
   confirmDelete(device: Device): void {
@@ -58,11 +74,20 @@ export class DeviceListComponent implements OnInit {
   onDeleteConfirmed(): void {
     const device = this.deviceToDelete();
     if (!device) return;
-    this.deviceService.delete(device.id).subscribe({
-      next: () => {
-        this.toast.success(`"${device.name}" deleted successfully.`);
-        this.showDeleteDialog.set(false);
-      },
-    });
+    if (this.authService.isAdmin()) {
+      this.deviceService.delete(device.id).subscribe({
+        next: () => {
+          this.toast.success(`"${device.name}" deleted successfully.`);
+          this.showDeleteDialog.set(false);
+        },
+      });
+    } else {
+      this.deviceService.selfUnassign(device.id).subscribe({
+        next: () => {
+          this.toast.success(`"${device.name}" returned successfully.`);
+          this.showDeleteDialog.set(false);
+        },
+      });
+    }
   }
 }
